requirements.txt
Mezzanine==4.1.0
django-mathfilters
django-multiselectfield
django-paypal
